requirements.txt
records
click
PyYAML
schema
pandas